backend/requirements.txt
fastapi
uvicorn[standard]
sse-starlette
pydantic>=2.0
litellm
python-dotenv
psycopg2-binary
perplexityai
exa-py
